refactor(gtm): split account-selection into folder, guard auto-select
Break account-selection.js into account-selection/ with dedicated
notice components (none/single/multiple accounts) instead of one
monolithic file.

Guard the auto-select effect with hasFinishedResolution - without it,
existingAccounts could transiently report length 1 before resolution
finished, auto-selecting the wrong account.
